requirements.txt
# Core dependencies
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
pillow>=8.2.0
tqdm>=4.61.2

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Image processing and augmentation
opencv-python>=4.5.3
albumentations>=1.3.0

# Model architectures
segmentation-models-pytorch>=0.3.0
timm>=0.9.0

# Metrics and evaluation
torchmetrics>=0.11.0

# Experiment tracking
wandb>=0.15.0
# mlflow>=2.3.0  # Optional alternative

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# CLI tools
typer>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
joblib>=1.2.0